pyproject.toml
[project]
name = "ml-ops"
version = "0.1.0"
description = ""
authors = [
    {name = "Mykhailo Maslov",email = "maslov.mikhail200@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ultralytics (>=8.3.154,<9.0.0)",
    "wandb (>=0.20.1,<0.21.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "ray[data,serve,train,tune] (==2.46.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "evidently (>=0.7.7,<0.8.0); python_version>='3.10' and python_version<'4.0'",
    "clickhouse-driver (>=0.2.9,<0.3.0); python_version>='3.10' and python_version<'4.0'",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
